+++ src/Interfaces/product/ProductVariant.ts
@@ -1,4 +1,9 @@
 import Product from './Product';
+import { BasketLine } from '../basket/BasketLine';
+import OrderLine from '../order/OrderLine';
+import Stock from '../stock/Stock';
+import Review from '../review/Review';
+import Rating from '../review/Rating';
 
 export default interface ProductVariant {
     id?: string;
@@ -36,7 +41,21 @@ export default interface ProductVariant {
     otherHazards?: string;
     supplierName?: string;
     aggregateRating?: number;
-    totalReviews?: number;
     // belongs to
     product?: Product | null;
+    // has many
+    basketLines?: Array<BasketLine>;
+    orderLines?: Array<OrderLine>;
+    stockObjects?: Array<Stock>;
+    reviews?: Array<Review>;
+    accountedForReviews?: Array<Review>;
+    // accessors
+    totalSafetyStock?: number;
+    totalStockWithCustomer?: number;
+    totalStockUnits?: number;
+    totalOrders?: number;
+    totalReturnedDamaged?: number;
+    totalSold?: number;
+    totalReviews?: number;
+    ratings?: Rating;
 }