+++ src/Resources/Auth/Auth.ts
@@ -1,21 +1,24 @@
 import AbstractResource from '../AbstractResource';
+import { HttpClientConfig, HttpClientResponse } from '../../Interfaces/HttpClient';
 
 export default class Auth extends AbstractResource {
     initialise(): void {
         this.defaultAccess = 'guest';
     }
 
-    public forgotPassword: Function = this.createMethodFromPartialSpec({
+    public forgotPassword: (...args: HttpClientConfig) => HttpClientResponse = this.createMethodFromPartialSpec({
         method: 'POST',
         path: '/password/forgot',
     });
 
-    public resetPassword: Function = this.createMethodFromPartialSpec({
+    public resetPassword: (...args: HttpClientConfig) => HttpClientResponse = this.createMethodFromPartialSpec({
         method: 'POST',
         path: '/password/reset',
     });
 
-    public validateResetPasswordToken: Function = this.createMethodFromPartialSpec({
+    public validateResetPasswordToken: (
+        ...args: HttpClientConfig
+    ) => HttpClientResponse = this.createMethodFromPartialSpec({
         method: 'POST',
         path: '/password/validate',
     });